+++ python/lsst/ts/salobj/topics/remote_telemetry.py
@@ -21,6 +21,8 @@
 
 __all__ = ["RemoteTelemetry"]
 
+import warnings
+
 from . import read_topic
 
 
@@ -34,19 +36,27 @@ class RemoteTelemetry(read_topic.ReadTopic):
     name : `str`
         Telemetry topic name
     max_history : `int`, optional
-        Maximum number of historical items to read:
-
-        * 0 if none; strongly recommended for `RemoteCommand` & `AckCmdReader`
-        * 1 is recommended for events and telemetry
+        Deprecated because historical telemetry data is no longer available.
+        Must be 0 (or None, but please don't do that) if specified.
     queue_len : `int`
         Number of elements that can be queued for `get_oldest`.
     """
 
-    def __init__(self, salinfo, name, max_history=1, queue_len=100):
+    def __init__(self, salinfo, name, max_history=None, queue_len=100):
+        # TODO DM-26474: remove the max_history argument and this code block.
+        if max_history is not None:
+            if max_history == 0:
+                warnings.warn("max_history is deprecated", DeprecationWarning)
+            else:
+                raise ValueError(
+                    f"max_history={max_history} is deprecated "
+                    "and must be 0 (or None, but please don't do that) if specified"
+                )
+
         super().__init__(
             salinfo=salinfo,
             name=name,
             sal_prefix="",
-            max_history=max_history,
+            max_history=0,
             queue_len=queue_len,
         )